% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackage.R
\name{get_package_functions}
\alias{get_package_functions}
\title{Discover Functions in a Package Using LLM}
\usage{
get_package_functions(package_name, model_name = "llama3.2:3b-instruct-q4_K_M")
}
\arguments{
\item{package_name}{Character string containing the name of an R package}

\item{model_name}{Character string with the name of the Ollama model to use (default: "llama3.2:3b-instruct-q4_K_M")}
}
\value{
A character vector of function names
}
\description{
This function uses the llama3.2:3b model through Ollama to predict
which functions might be exported by a given R package.
}
